Split off parts of single pass compiler
